requirements.txt
ImageHash~=4.3.1
Pillow~=10.4.0
pip~=24.1.2
wheel~=0.43.0
setuptools~=71.0.3

tksheet~=7.2.12
packaging~=24.1
